Move ./cmd/tool commands to a package
