+++ cmd/tool/main.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/cli"
 	"github.com/spf13/cobra"
+	"github.com/tektoncd/operator/cmd/tool/commands"
 )
 
 func main() {
@@ -20,9 +21,9 @@ func main() {
 
 	ioStreams := cli.NewIOStreams()
 
-	cmd.AddCommand(BumpCommand(ioStreams))
-	cmd.AddCommand(CheckCommand(ioStreams))
-	cmd.AddCommand(ComponentVersionCommand(ioStreams))
+	cmd.AddCommand(commands.BumpCommand(ioStreams))
+	cmd.AddCommand(commands.CheckCommand(ioStreams))
+	cmd.AddCommand(commands.ComponentVersionCommand(ioStreams))
 
 	if err := cmd.Execute(); err != nil {
 		os.Exit(1)
